20240308-devops-simplicity/cmd/hello_test.go
package main

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/labstack/echo/v4"
	"github.com/stretchr/testify/assert"
)

func TestHelloHandler(t *testing.T) {
	e := echo.New()
	req := httptest.NewRequest(http.MethodGet, "/", nil)
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec)

	// Call the handler
	err := helloHandler(c)
	if err != nil {
		t.Errorf("helloHandler returned an error: %v", err)
	}

	// Check the response
	assert.Equal(t, http.StatusOK, rec.Code)
}

func TestHealthHandler(t *testing.T) {
	e := echo.New()
	req := httptest.NewRequest(http.MethodGet, "/health", nil)
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec)

	// Call the handler
	err := healthHandler(c)
	if err != nil {
		t.Errorf("healthHandler returned an error: %v", err)
	}

	// Check the response
	assert.Equal(t, http.StatusOK, rec.Code)
	assert.Equal(t, "healthy", rec.Body.String())
}
